+++ R/parsers.R
@@ -207,6 +207,7 @@ parse_mothur_tax_summary <- function(file = NULL, text = NULL, table = NULL) {
     # replace taxon names
     my_taxon_names <- output$map_data_(output$taxon_ids(),
                                        output$get_data("taxon")[[1]])
+    my_taxon_names <- as.character(my_taxon_names)
     output$taxa <- stats::setNames(lapply(seq_len(length(output$taxa)),
                                           function(i) {
                                             my_taxon <- output$taxa[[i]]
@@ -464,7 +465,7 @@ parse_unite_general <- function(input = NULL, file = NULL, include_seqs = TRUE) 
                                            unite_type = "info",
                                            tax_string = "class"),
                                    class_regex = "^(.*)__(.*)$",
-                                   class_key = c(unite_rank = "info",
+                                   class_key = c(unite_rank = "taxon_rank",
                                                  name = "taxon_name"),
                                    class_sep = ";")