+++ NEWS.md
@@ -1,5 +1,9 @@
 # News 
 
+## Current
+
+* Added parsing of taxon ranks to the UNITE parser
+
 ## metacoder 0.3.3
 
 * Fixed a bug in `calc_n_samples` and `calc_prop_samples` that caused an error when group names had spaces in them.